Move CLI logic to medusa.py
